+++ lib/maze/client/appium/bb_client.rb
@@ -27,27 +27,36 @@ module Maze
         def device_capabilities
           config = Maze.config
           capabilities = {
-            'disabledAnimations' => 'true',
-            'noReset' => 'true',
+            'noReset' => true,
             'bitbar:options' => {
               # Some capabilities probably belong in the top level
               # of the hash, but BitBar picks them up from here.
               'apiKey' => config.access_key,
               'app' => config.app,
-              'testrun' => "#{config.os} #{config.os_version}",
               'findDevice' => false,
               'testTimeout' => 7200,
               'newCommandTimeout' => 0
             }
           }
           capabilities['appiumVersion'] = config.appium_version unless config.appium_version.nil?
+          capabilities.deep_merge! dashboard_capabilities
           capabilities.deep_merge! BitBarDevices.get_available_device(config.device)
           capabilities.deep_merge! JSON.parse(config.capabilities_option)
           capabilities
         end
 
-        def log_session_info
-          # Not yet implemented
+        def log_run_intro
+          # Nothing to log at the start
+        end
+
+        def log_run_outro
+          api_client = BitBarApiClient.new
+
+          $logger.info 'Appium session(s) created:'
+          @session_ids.each do |id|
+            link = api_client.get_device_session_ui_link(id)
+            $logger.info Maze::LogUtil.linkify(link, "BitBar session: #{id}") if link
+          end
         end
 
         def stop_session
@@ -55,6 +64,37 @@ module Maze
           if Maze.config.start_tunnel
             Maze::Client::BitBarClientUtils.stop_local_tunnel
           end
+        end
+
+        # Determines capabilities used to organise sessions in the BitBar dashboard.
+        #
+        # @return [Hash] A hash containing the capabilities.
+        def dashboard_capabilities
+          # Attempt to use the current git repo as the project name
+          output, status = Maze::Runner.run_command('git rev-parse --show-toplevel')
+          if status == 0
+            project = File.basename(output[0].strip)
+          else
+            if ENV['BUILDKITE']
+              project = ENV['BUILDKITE_PIPELINE_SLUG']
+            else
+              $logger.warn 'Unable to determine project name, consider running Maze Runner from within a Git repository'
+              project = 'Unknown'
+            end
+          end
+
+          if ENV['BUILDKITE']
+            test_run = "#{ENV['BUILDKITE_BUILD_NUMBER']} - #{ENV['BUILDKITE_LABEL']}"
+          else
+            test_run = Maze.run_uuid
+          end
+
+          {
+            'bitbar:options' => {
+              bitbar_project: project,
+              bitbar_testrun: test_run
+            }
+          }
         end
       end
     end