+++ lib/maze/client/appium/bs_client.rb
@@ -50,20 +50,16 @@ module Maze
           capabilities
         end
 
-        def log_session_info
-          if Maze.config.device || Maze.config.browser
-            # Log device id
-            udid = Maze.driver.session_capabilities['udid']
-            $logger.info "Running on device: #{udid}" unless udid.nil?
+        def log_run_intro
+          # Log a link to the BrowserStack session search dashboard
+          url = "https://app-automate.browserstack.com/dashboard/v2/search?query=#{Maze.run_uuid}&type=builds"
+          $logger.info Maze::LogUtil.linkify(url, 'BrowserStack session(s)')
+        end
 
-            # Log a link to the BrowserStack session search dashboard
-            url = "https://app-automate.browserstack.com/dashboard/v2/search?query=#{Maze.run_uuid}&type=builds"
-            if ENV['BUILDKITE']
-              $logger.info Maze::LogUtil.linkify(url, 'BrowserStack session(s)')
-            else
-              $logger.info "BrowserStack session(s): #{url}"
-            end
-          end
+        def log_run_outro
+          $logger.info 'Appium session(s) created:'
+          @session_ids.each { |id| $logger.info "  #{id}" }
+          log_run_intro
         end
 
         def stop_session